+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 ## 0.2.0 [unreleased]
 
+### Features
+
+1. [#30](https://github.com/InfluxCommunity/influxdb3-go/pull/30): Add custom HTTP headers support
+
+### Breaking Changes
+
+1. [#31](https://github.com/InfluxCommunity/influxdb3-go/pull/31): Changed package to `influxdb3`.
+Renamed config types and some options.
+
 ## 0.1.0 [2023-06-09]
 
 - initial release of new client version
 - write using v2 api
 - query using flightSQL
-- query using influxQl
+- query using influxQL